# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Yercaud Business Directory API
  version: 0.1.0
  description: >
    Backend contract for the public frontend and Admin panel. Auth is
    httpOnly session cookies (ADR: session cookies over JWT). See
    /CONTEXT.md and /docs/adr/ for the domain model and design decisions
    this contract implements.
servers:
  - url: http://localhost:4000/v1
    description: Local dev
security:
  - cookieAuth: []
...
paths:
...
  /businesses:
    get:
      tags: [Businesses]
      summary: List businesses (Super Admin sees all; Business Owner sees own)
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/ApprovalStatus' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Business' }
    post:
      tags: [Businesses]
      summary: Apply as a Business Owner (FR161) — starts as 'pending'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BusinessInput' }
      responses:
        '201':
          description: Created, pending Super Admin approval
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Business' }
  /businesses/{businessId}:
    parameters:
      - $ref: '#/components/parameters/BusinessId'
    get:
      tags: [Businesses]
      security: []
      summary: Get a Business — public once approved, owner/admin only before that
      description: >-
        An approved Business is readable by anyone: the public Listing page shows
        its address, email and social links, and a directory whose businesses
        can't be read is not a directory. (Phone and website were already public,
        riding along on ListingSummary — the old rule wasn't self-consistent.)

        Anything not yet approved keeps the previous rule exactly: 401 when not
        signed in, 403 when signed in without ownership or Businesses:view. Not
        404 — Phase 1 chose that distinction deliberately, and Business ids are
        UUIDs, so the enumeration argument that justifies 404 for Listings
        doesn't apply.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Business' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Businesses]
      summary: Update a business (owner or Super Admin) — applies immediately, ADR 0005
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BusinessInput' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Business' }
  /businesses/{businessId}/approve:
    parameters: [{ $ref: '#/components/parameters/BusinessId' }]
    post:
      tags: [Businesses]
      summary: Approve a pending business (Super Admin)
      responses:
        '200': { description: Approved }
  /businesses/{businessId}/reject:
    parameters: [{ $ref: '#/components/parameters/BusinessId' }]
    post:
      tags: [Businesses]
      summary: Reject a pending business (Super Admin)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties: { reason: { type: string } }
      responses:
        '200': { description: Rejected }
  /businesses/{businessId}/suspend:
    parameters: [{ $ref: '#/components/parameters/BusinessId' }]
    post:
      tags: [Businesses]
      summary: Suspend a business for policy violation (Super Admin)
      responses:
        '200': { description: Suspended }
  /businesses/{businessId}/archive:
    parameters: [{ $ref: '#/components/parameters/BusinessId' }]
    post:
      tags: [Businesses]
      summary: Archive a business — history kept, hidden from public site (ADR 0010)
      responses:
        '200': { description: Archived }
...
  /blog-posts:
    get:
      tags: [Blog]
      security: []
      summary: List/search blog posts, newest-first (public sees 'published' only)
      parameters:
        - name: q
          in: query
          description: Case-insensitive substring match over title, excerpt, and body.
          schema: { type: string }
        - name: category
          in: query
          description: >-
            The blog category's slug (unlike GET /faqs' `category`, which takes
            a UUID — FAQ categories have no slug).
          schema: { type: string }
        - name: status
          in: query
          description: Admin callers only
          schema: { type: string, enum: [draft, pending, published] }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/BlogPostSummary' }
    post:
      tags: [Blog]
      summary: Create a blog post (Super Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BlogPostInput' }
      responses:
        '201': { description: Created as draft }
  /blog-posts/by-slug/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Blog]
      security: []
      summary: Get a blog post by its slug (the public site's URL shape)
      description: >-
        The public site addresses Blog Posts by slug — readable, keyword-bearing,
        and what /blog/$slug links to. Admin keeps using ids. A separate path
        rather than overloading {postId}, so a slug that happens to look like a
        UUID is never ambiguous. Same visibility rule as fetching by id: an
        unpublished post 404s to anyone without Content:view.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BlogPost' }
        '404': { $ref: '#/components/responses/NotFound' }
  /blog-posts/{postId}:
    parameters:
      - $ref: '#/components/parameters/PostId'
    get:
      tags: [Blog]
      security: []
      summary: Get a blog post
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BlogPost' }
    patch:
      tags: [Blog]
      summary: Update a blog post (Super Admin) — partial; send only what changes
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BlogPostUpdate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BlogPost' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Blog]
      summary: Delete a blog post (Super Admin) — cascades to its comments and ratings
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /blog-posts/{postId}/publish:
    parameters: [{ $ref: '#/components/parameters/PostId' }]
    post:
      tags: [Blog]
      summary: >-
        Publish a blog post (Super Admin). Stamps publishedAt on the first
        publish only — re-publishing an edited post must not reorder the
        reader-facing feed, and is a no-op if it's already published.
      responses:
        '200':
          description: Published
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BlogPost' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /blog-posts/{postId}/duplicate:
    parameters: [{ $ref: '#/components/parameters/PostId' }]
    post:
      tags: [Blog]
      summary: >-
        Duplicate a blog post into a new draft (Super Admin) — title suffixed
        " (Copy)", re-slugged, status/publishedAt reset. Curated links
        (relatedPostIds/placeListingIds) aren't copied.
      responses:
        '201':
          description: Created as draft
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BlogPost' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_id
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    BusinessId:
      name: businessId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    PostId:
      name: postId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Forbidden:
      description: Not permitted (RBAC)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
    ApprovalStatus:
      type: string
      enum: [pending, approved, rejected, suspended, archived]
...
    Business:
      type: object
      required: [id, ownerId, name, description, contactPhone, contactEmail, website, address, socialLinks, logoUrl, coverUrl, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        ownerId: { type: string, format: uuid }
        name: { type: string }
        description: { type: string, nullable: true }
        contactPhone: { type: string, nullable: true }
        contactEmail: { type: string, nullable: true }
        website: { type: string, nullable: true }
        address: { type: string, nullable: true }
        socialLinks: { type: object, additionalProperties: { type: string } }
        logoUrl: { type: string, nullable: true }
        coverUrl: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/ApprovalStatus' }
        rejectionReason: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    BusinessInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
        contactPhone: { type: string }
        contactEmail: { type: string }
        website: { type: string }
        address: { type: string }
        socialLinks: { type: object, additionalProperties: { type: string } }
        logoUrl: { type: string }
        coverUrl: { type: string }
...
    BlogPostSummary:
      type: object
      # Declared required because the response always carries them — same
      # reasoning as ListingSummary above (Phase 10): without this every
      # generated field is `| undefined`, which defeats generated types.
      required:
        - id
        - title
        - slug
        - excerpt
        - coverImage
        - categoryId
        - authorId
        - authorName
        - authorAvatarUrl
        - authorBio
        - readingTimeMinutes
        - status
        - publishedAt
        - createdAt
        - viewCount
        - tags
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        slug: { type: string }
        excerpt: { type: string, nullable: true }
        coverImage: { type: string, nullable: true }
        categoryId: { type: string, format: uuid }
        authorId: { type: string, format: uuid }
        # author_id is NOT NULL, so these are always populated — unlike
        # Review's authorName, which is optional because that module has a
        # second query variant that doesn't join users at all.
        authorName: { type: string }
        authorAvatarUrl: { type: string, nullable: true }
        authorBio: { type: string, nullable: true }
        readingTimeMinutes: { type: integer, nullable: true }
        status: { type: string, enum: [draft, pending, published] }
        publishedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        viewCount: { type: integer }
        tags: { type: array, items: { type: string } }
    BlogPost:
      allOf:
        - $ref: '#/components/schemas/BlogPostSummary'
        - type: object
          # Declared required because the response always carries them — same
          # reasoning as BlogPostSummary/BlogComment above (toBlogPost always
          # sets every one of these, nullable fields included, to null rather
          # than omitting them).
          required:
            - body
            - seoTitle
            - seoDescription
            - ogImage
            - twitterCard
            - relatedPostIds
            - placeListingIds
          properties:
            body: { type: string }
            seoTitle: { type: string, nullable: true }
            seoDescription: { type: string, nullable: true }
            # FR189, closed in Phase 10. Phase 9 exposed these on Listings, where
            # the columns already existed unused, and left Blog Posts out because
            # they needed a migration — noting the asymmetry it created. A Blog
            # Post is the thing most likely to be shared into WhatsApp, so the
            # asymmetry was backwards.
            ogImage: { type: string, nullable: true }
            twitterCard:
              type: string
              nullable: true
              enum: [summary, summary_large_image, app, player]
            relatedPostIds: { type: array, items: { type: string, format: uuid } }
            placeListingIds: { type: array, items: { type: string, format: uuid } }
    # `slug`, `readingTimeMinutes`, `authorId`, `status`, and `publishedAt` are
    # all absent by design: the first two are derived server-side from title
    # and body, authorId comes from the session, and status only moves via
    # PATCH or /publish. A new post always starts as 'draft'.
    BlogPostInput:
      type: object
      required: [categoryId, title, body]
      properties:
        categoryId: { type: string, format: uuid }
        title: { type: string }
        excerpt: { type: string, nullable: true }
        body: { type: string }
        coverImage: { type: string, nullable: true }
        seoTitle: { type: string, nullable: true }
        seoDescription: { type: string, nullable: true }
        # A URL to an already-hosted image, not an upload — same as a Listing's.
        ogImage: { type: string, nullable: true }
        twitterCard:
          type: string
          nullable: true
          enum: [summary, summary_large_image, app, player]
        # A post may not relate to itself (400).
        relatedPostIds: { type: array, items: { type: string, format: uuid } }
        # Every id must be an approved Listing (400) — a published article must
        # not link somewhere readers can't go (FR114).
        placeListingIds: { type: array, items: { type: string, format: uuid } }
        # Freeform tag names — resolved to blog_tags rows via get-or-create.
        tags: { type: array, items: { type: string } }
    BlogPostUpdate:
      type: object
      properties:
        categoryId: { type: string, format: uuid }
        title: { type: string }
        excerpt: { type: string, nullable: true }
        body: { type: string }
        coverImage: { type: string, nullable: true }
        seoTitle: { type: string, nullable: true }
        seoDescription: { type: string, nullable: true }
        relatedPostIds: { type: array, items: { type: string, format: uuid } }
        placeListingIds: { type: array, items: { type: string, format: uuid } }
        tags: { type: array, items: { type: string } }
        status: { type: string, enum: [draft, pending, published] }
        # Scheduling: an explicit future date holds the post out of the
        # public feed until it passes (published_at <= NOW() check) — only
        # meaningful when status is (becoming) 'published'.
        publishedAt: { type: string, format: date-time }
...